backend/requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
python-dotenv==1.1.1
djangorestframework==3.15.2
django-cors-headers==4.5.0
gunicorn==22.0.0
